pyproject.toml
[tool.poetry]
name = "jmarci-de1-v2-4-1"
version = "0.1.1"
description = "Capstone project"
authors = ["Julius <j.marciulynas@gmail.com>"]
readme = "README.md"
packages = [
    { include = "jmarci_de1" },
    { include = "utils" }
]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
sqlalchemy-utils = "^0.41.2"
fastapi = "^0.111.0"
python-decouple = "^3.8"
kaggle = "^1.6.14"
pymysql = "^1.1.1"
pandas = "^2.2.2"
datetime = "^5.5"
cryptography = "^42.0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
